// src-old/proto/js/Agent_pb.d.ts
    export type AsObject = {
    }
// ...
export class LinkInfoIdentityMessage extends jspb.Message { 
    getType(): string;
    setType(value: string): LinkInfoIdentityMessage;

    getNode(): string;
    setNode(value: string): LinkInfoIdentityMessage;

    getIdentity(): string;
    setIdentity(value: string): LinkInfoIdentityMessage;

    getProvider(): string;
    setProvider(value: string): LinkInfoIdentityMessage;

    getDateissued(): string;
    setDateissued(value: string): LinkInfoIdentityMessage;

    getSignature(): string;
    setSignature(value: string): LinkInfoIdentityMessage;

    getKey(): string;
    setKey(value: string): LinkInfoIdentityMessage;

    getUrl(): string;
    setUrl(value: string): LinkInfoIdentityMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): LinkInfoIdentityMessage.AsObject;
    static toObject(includeInstance: boolean, msg: LinkInfoIdentityMessage): LinkInfoIdentityMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: LinkInfoIdentityMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): LinkInfoIdentityMessage;
    static deserializeBinaryFromReader(message: LinkInfoIdentityMessage, reader: jspb.BinaryReader): LinkInfoIdentityMessage;
}

export namespace LinkInfoIdentityMessage {
    export type AsObject = {
        type: string,
        node: string,
        identity: string,
        provider: string,
        dateissued: string,
        signature: string,
        key: string,
        url: string,
    }
}

export class LinkInfoNodeMessage extends jspb.Message { 
    getType(): string;
    setType(value: string): LinkInfoNodeMessage;

    getNode1(): string;
    setNode1(value: string): LinkInfoNodeMessage;

    getNode2(): string;
    setNode2(value: string): LinkInfoNodeMessage;

    getDateissued(): string;
    setDateissued(value: string): LinkInfoNodeMessage;

    getSignature(): string;
    setSignature(value: string): LinkInfoNodeMessage;

    getKey(): string;
    setKey(value: string): LinkInfoNodeMessage;

    getUrl(): string;
    setUrl(value: string): LinkInfoNodeMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): LinkInfoNodeMessage.AsObject;
    static toObject(includeInstance: boolean, msg: LinkInfoNodeMessage): LinkInfoNodeMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: LinkInfoNodeMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): LinkInfoNodeMessage;
    static deserializeBinaryFromReader(message: LinkInfoNodeMessage, reader: jspb.BinaryReader): LinkInfoNodeMessage;
}

export namespace LinkInfoNodeMessage {
    export type AsObject = {
        type: string,
        node1: string,
        node2: string,
        dateissued: string,
        signature: string,
        key: string,
        url: string,
    }
}

export class LinkInfoMessage extends jspb.Message { 

    hasLinkInfoIdentity(): boolean;
    clearLinkInfoIdentity(): void;
    getLinkInfoIdentity(): LinkInfoIdentityMessage | undefined;
    setLinkInfoIdentity(value?: LinkInfoIdentityMessage): LinkInfoMessage;


    hasLinkInfoNode(): boolean;
    clearLinkInfoNode(): void;
    getLinkInfoNode(): LinkInfoNodeMessage | undefined;
    setLinkInfoNode(value?: LinkInfoNodeMessage): LinkInfoMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): LinkInfoMessage.AsObject;
    static toObject(includeInstance: boolean, msg: LinkInfoMessage): LinkInfoMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: LinkInfoMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): LinkInfoMessage;
    static deserializeBinaryFromReader(message: LinkInfoMessage, reader: jspb.BinaryReader): LinkInfoMessage;
}

export namespace LinkInfoMessage {
    export type AsObject = {
        linkInfoIdentity?: LinkInfoIdentityMessage.AsObject,
        linkInfoNode?: LinkInfoNodeMessage.AsObject,
    }
}

export class GestaltMatrixEdgeMessage extends jspb.Message { 

    getPairsMap(): jspb.Map<string, LinkInfoMessage>;
    clearPairsMap(): void;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): GestaltMatrixEdgeMessage.AsObject;
    static toObject(includeInstance: boolean, msg: GestaltMatrixEdgeMessage): GestaltMatrixEdgeMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: GestaltMatrixEdgeMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): GestaltMatrixEdgeMessage;
    static deserializeBinaryFromReader(message: GestaltMatrixEdgeMessage, reader: jspb.BinaryReader): GestaltMatrixEdgeMessage;
}

export namespace GestaltMatrixEdgeMessage {
    export type AsObject = {

        pairsMap: Array<[string, LinkInfoMessage.AsObject]>,
    }
}

export class GestaltNodeMessage extends jspb.Message { 
    getId(): string;
    setId(value: string): GestaltNodeMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): GestaltNodeMessage.AsObject;
    static toObject(includeInstance: boolean, msg: GestaltNodeMessage): GestaltNodeMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: GestaltNodeMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): GestaltNodeMessage;
    static deserializeBinaryFromReader(message: GestaltNodeMessage, reader: jspb.BinaryReader): GestaltNodeMessage;
}

export namespace GestaltNodeMessage {
    export type AsObject = {
        id: string,
    }
}

export class GestaltMessage extends jspb.Message { 

    getGestaltMatrixMap(): jspb.Map<string, GestaltMatrixEdgeMessage>;
    clearGestaltMatrixMap(): void;


    getGestaltNodesMap(): jspb.Map<string, GestaltNodeMessage>;
    clearGestaltNodesMap(): void;


    getIdentitiesMap(): jspb.Map<string, IdentityInfoMessage>;
    clearIdentitiesMap(): void;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): GestaltMessage.AsObject;
    static toObject(includeInstance: boolean, msg: GestaltMessage): GestaltMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: GestaltMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): GestaltMessage;
    static deserializeBinaryFromReader(message: GestaltMessage, reader: jspb.BinaryReader): GestaltMessage;
}

export namespace GestaltMessage {
    export type AsObject = {

        gestaltMatrixMap: Array<[string, GestaltMatrixEdgeMessage.AsObject]>,

        gestaltNodesMap: Array<[string, GestaltNodeMessage.AsObject]>,

        identitiesMap: Array<[string, IdentityInfoMessage.AsObject]>,
    }
}
// ...
export class IdentityInfoMessage extends jspb.Message { 
    getKey(): string;
    setKey(value: string): IdentityInfoMessage;

    getProvider(): string;
    setProvider(value: string): IdentityInfoMessage;

    getName(): string;
    setName(value: string): IdentityInfoMessage;

    getEmail(): string;
    setEmail(value: string): IdentityInfoMessage;

    getUrl(): string;
    setUrl(value: string): IdentityInfoMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): IdentityInfoMessage.AsObject;
    static toObject(includeInstance: boolean, msg: IdentityInfoMessage): IdentityInfoMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: IdentityInfoMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): IdentityInfoMessage;
    static deserializeBinaryFromReader(message: IdentityInfoMessage, reader: jspb.BinaryReader): IdentityInfoMessage;
}

export namespace IdentityInfoMessage {
    export type AsObject = {
        key: string,
        provider: string,
        name: string,
        email: string,
        url: string,
    }
}